# perf: account alias lookup