### What I did

In [7]: %time accounts.load("laughing-sepolia-safe")
CPU times: user 566 µs, sys: 497 µs, total: 1.06 ms
Wall time: 729 µs
Out[7]: <SafeAccount 0x872Df5f6CF2C2Eaa9e7511AB079D263AB8fdc79f>

In [8]: %time accounts.load("laughing-sepolia-safe")
CPU times: user 7 µs, sys: 1e+03 ns, total: 8 µs
Wall time: 9.78 µs
Out[8]: <SafeAccount 0x872Df5f6CF2C2Eaa9e7511AB079D263AB8fdc79f>

**Note:** Some accounts plugins a bit slower than others on first load.

### How I did it

<!-- Discuss the thought process behind the change -->

### How to verify it

<!-- Discuss any methods that should be used to verify the change -->

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated